move SendCache class into it's own file, not exported in index.ts
